requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.2.1
iniconfig==2.0.0
Markdown==3.6
packaging==24.0
pluggy==1.5.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyotp==2.9.0
pytest==8.2.0
pytest-django==4.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
